requirements.txt
fastapi
uvicorn
sqlalchemy
passlib[bcrypt]
python-multipart
python-dotenv
psycopg2-binary
PyJWT 